Refactorizacion del codigo dejando el archivo data con funciones puras y main con la manipulacion del DOM
