+++ test/data.spec.js
@@ -1,23 +1,30 @@
-import { example, anotherExample } from '../src/data.js';
-
-
-describe('example', () => {
+// import { example, anotherExample } from '../src/data.js';
+import countriesDataAdmin from '../src/data.js';
+//crear una const que sera igual a un objeto 
+describe('generateSection', () => {
   it('is a function', () => {
-    expect(typeof example).toBe('function');
-  });
-
-  it('returns `example`', () => {
-    expect(example()).toBe('example');
-  });
+    expect(typeof countriesDataAdmin).toBe('object')
+  })
 });
 
 
-describe('anotherExample', () => {
-  it('is a function', () => {
-    expect(typeof anotherExample).toBe('function');
-  });
+// describe('example', () => {
+//   it('is a function', () => {
+//     expect(typeof example).toBe('function');
+//   });
 
-  it('returns `anotherExample`', () => {
-    expect(anotherExample()).toBe('OMG');
-  });
-});
+//   it('returns `example`', () => {
+//     expect(example()).toBe('example');
+//   });
+// });
+
+
+// describe('anotherExample', () => {
+//   it('is a function', () => {
+//     expect(typeof anotherExample).toBe('function');
+//   });
+
+//   it('returns `anotherExample`', () => {
+//     expect(anotherExample()).toBe('OMG');
+//   });
+// });
